bug fixes with conversion to integers on non-Linux machine. Also updated all integer conversions of NaR to 0.

+++ source/p16_to_i32.c
@@ -73,7 +73,7 @@ int_fast32_t p16_to_i32( posit16_t pA ){
 		}
 		uiA <<= 1;                           // Skip over termination bit, which is 0.
 		if (0x2000 & uiA) scale++;           // If exponent is 1, increment the scale.
-		iZ = (uiA | 0x2000) << 17;         // Left-justify fraction in 32-bit result (one left bit padding)
+		iZ = ((uint32_t)uiA | 0x2000) << 17;         // Left-justify fraction in 32-bit result (one left bit padding)
 		mask = 0x40000000 >> scale;          // Point to the last bit of the integer part.
 
 		bitLast = (iZ & mask);               // Extract the bit, without shifting it.
@@ -88,7 +88,7 @@ int_fast32_t p16_to_i32( posit16_t pA ){
 			if (bitLast | tmp) iZ += (mask << 1);
 		}
 
-		iZ = iZ >> (30 - scale);             // Right-justify the integer.
+		iZ = (uint32_t)iZ >> (30 - scale);             // Right-justify the integer.
 	}
 
 	if (sign) iZ = -iZ;                      // Apply the sign of the input.
